+++ OnlySnarf/settings.py
@@ -27,7 +27,6 @@ class Settings:
         self.ACTION = "upload"
         self.AMOUNT = 0
         self.MONTHS = 0
-        self.CHOICE = None
         self.PRICE = 0
         self.METHOD = "random"
         self.CRON = False
@@ -38,6 +37,8 @@ class Settings:
         self.SCHEDULE = None
         self.DATE = None
         self.TIME = None
+        self.DURATION = None
+        self.QUESTIONS = []
         # -backup
         # backup uploaded content to "posted" folder
         self.BACKUP = False
@@ -186,7 +187,7 @@ class Settings:
             sys.argv[i] = sys.argv[i][1:] # remove - in front
             truths_ = ["BACKUP","CREATE_DRIVE","DEBUG","DEBUG_DELAY","DELETE_GOOGLE","FORCE_DELETE","FORCE_UPLOAD","FORCE_REDUCTION","PREFER_LOCAL","SAVE_USERS","SHOW_WINDOW","SKIP_DELETE","SKIP_DOWNLOAD","SKIP_REDUCE","SKIP_REPAIR","SKIP_UPLOAD","TWEETING","VERBOSE","THUMBNAILING_PREVIEW"]
             falses_ = []
-            nexts_ = ["DATE","TIME","SCHEDULE","EXPIRES","USERS_FAVORITE","CRON","METHOD","PRICE","CHOICE","AMOUNT","MONTHS","ACTION","CRON_USER","LOCAL","IMAGE","IMAGE_UPLOAD_LIMIT","IMAGE_UPLOAD_MAX","TYPE","TEXT","USER","DRIVE_PATH","GOOGLE_PATH","MOUNT_PATH","USERS_PATH","USERNAME","PASSWORD","USER_ID"]
+            nexts_ = ["DURATION","QUESTIONS","DATE","TIME","SCHEDULE","EXPIRES","USERS_FAVORITE","CRON","METHOD","PRICE","AMOUNT","MONTHS","ACTION","CRON_USER","LOCAL","IMAGE","IMAGE_UPLOAD_LIMIT","IMAGE_UPLOAD_MAX","TYPE","TEXT","USER","DRIVE_PATH","GOOGLE_PATH","MOUNT_PATH","USERS_PATH","USERNAME","PASSWORD","USER_ID"]
             j = 0
             while j < len(truths_):
                 if str(truths_[j]).upper() in str(sys.argv[i]).upper().replace("-","_"):
@@ -263,6 +264,14 @@ class Settings:
             self.maybePrint("Error: Missing Local Path")
             return False
         return self.LOCAL
+
+    def getPoll(self):
+        if isinstance(self.QUESTIONS, str): self.QUESTIONS = self.QUESTIONS.split(",")
+        poll = None
+        duration = self.DURATION or ""
+        questions = self.QUESTIONS or []
+        poll = {"period":duration,"questions":questions}
+        return poll
 
     def getSchedule(self):
         if str(self.SCHEDULE) != "None": return self.SCHEDULE