+++ CHANGELOG.md
@@ -219,8 +219,11 @@
   - add email|Twitter functionality for sending trial link; add clipboard function to copy link
 
 ### High Priority  
- 
-  -> Cron & Bot functionality that checks user messages for bot commands
+  - upload a gallery to a message
+
+  -> Cron
+  - bot functionality that checks user messages for bot commands
+  - ability to download images to upload / send later
 
   -> Twitter
   - tweet reminders